feat: override equals method rather than overload
Every user-defined class is a subclass of the Object class in java.
Therefore, it inherits the equals method from it. This method receives
an object of type Object as an argument. So, to override this method in
the new class we need to write the method signature equally to how it
appears in the Object class definition. From there, we would add an
extra line to type cast the argument of type Object to one of type Toy.